server/app/requirements.txt
aiobotocore==2.25.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.13.0
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.31.0
attrs==25.4.0
bcrypt==4.3.0
botocore==1.40.70
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
decorator==4.4.2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.122.0
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.3
moviepy==1.0.3
multidict==6.7.0
numpy==2.3.5
passlib==1.7.4
pillow==11.3.0
proglog==0.1.12
propcache==0.4.1
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wrapt==1.17.3
yarl==1.22.0
